# Conflicts:
#	TODO

+++ test/xmlsuite/txttoxml.py
@@ -2,7 +2,7 @@
 #*****************************************************************************/
 #*  LibreDWG - free implementation of the DWG file format                    */
 #*                                                                           */
-#*  Copyright (C) 2011 Free Software Foundation, Inc.                        */
+#*  Copyright (C) 2011, 2018 Free Software Foundation, Inc.                  */
 #*  Copyright (C) 2011 Guruprasad Rane                                       */
 #*                                                                           */
 #*  This library is free software, licensed under the terms of the GNU       */
@@ -67,7 +67,7 @@ if Extension[-1] == 'txt':
 else:
 	XMLFileName = OutDir + "/" + DwgTxtFileName + "xml"
 
-if os.path.exists(DwgTxtFileName):
+if os.path.exists(DwgTxtFileName) and os.path.exists(OutDir):
         FR = open(DwgTxtFileName,'r')
         FW = open(XMLFileName,'w')
         for Line in FR:
